[wmco] Drop all reference to WMCB.exe & submodule
The `windows-machine-config-bootstrapper` is retired in favor of WICD.
WMCB is no longer shipped as part of the WMCO payload.
Ran `git rm --force windows-machine-config-bootstrapper/`

The `wget-ignore-cert.ps1` script has also been dropped since this was only used
by WMCB to get ignition from the MCS worker ignition endpoint. We now get
ignition from the worker MachineConfig resource, which is why the worker
ignition endpoint has also been dropped from the internal WMCO cache.
