+++ pkg/csr/csr.go
@@ -306,9 +306,8 @@ func matchesHostname(nodeName string, windowsInstances []*instance.Info,
 
 // findHostName returns the actual host name of the instance by running the 'hostname' command
 func findHostName(instanceInfo *instance.Info, instanceSigner ssh.Signer) (string, error) {
-	// We don't need to pass the platform type here because this parameter is required for WMCB only.
-	// Here we just execute the "hostname" command.
-	win, err := windows.New("", "", "", instanceInfo, instanceSigner, "")
+	// We don't need to pass most args here as we just need to be able to run commands on the instance.
+	win, err := windows.New("", "", instanceInfo, instanceSigner)
 	if err != nil {
 		return "", errors.Wrap(err, "error instantiating Windows instance")
 	}
