+++ pkg/nodeconfig/init.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"strings"
 
+	clientset "github.com/openshift/client-go/config/clientset/versioned"
 	"github.com/pkg/errors"
 	core "k8s.io/api/core/v1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -19,14 +20,9 @@ import (
 // cache holds the information of the nodeConfig that is invariant for multiple reconciliation cycles. We'll use this
 // information when we don't want to get the information from the global context coming from reconciler
 // but to have something at nodeConfig package locally which will be passed onto other structs. There is no need to
-// invalidate this cache as of now, since the only entry in this workerIgnitionEndPoint which will be immutable. If
-// someone wants to change it, they've to restart the operator which will invalidate the cache automatically.
-// Note: It is ok to remove this struct in future, if we don't want to continue. As of now, I can think of only
-// worker ignition endpoint being part of this struct.
+// invalidate this cache as of now, since if someone wants to change any of the fields, they've to restart the operator
+// which will invalidate the cache automatically.
 type cache struct {
-	// workerIgnitionEndpoint is the Machine Config Server(MCS) endpoint from which we can download the
-	// the OpenShift worker ignition file.
-	workerIgnitionEndPoint string
 	// apiServerEndpoint is the address which clients can interact with the API server through
 	apiServerEndpoint string
 	// credentials holds a certificate and token needed to interact with the API server
@@ -46,18 +42,35 @@ func init() {
 		log.Error(err, "unable to find kube api server endpoint")
 		return
 	}
-	clusterAddress, err := getClusterAddr(kubeAPIServerEndpoint)
-	if err != nil {
-		log.Error(err, "error getting cluster address")
-		return
-	}
 	// populate the cache
 	nodeConfigCache.apiServerEndpoint = kubeAPIServerEndpoint
-	nodeConfigCache.workerIgnitionEndPoint = "https://" + clusterAddress + ":22623/config/worker"
 	nodeConfigCache.credentials, err = getWICDCredentials()
 	if err != nil {
 		log.Error(err, "unable to get WICD service account credentials")
 	}
+}
+
+// discoverKubeAPIServerEndpoint discovers the kubernetes api server endpoint
+func discoverKubeAPIServerEndpoint() (string, error) {
+	cfg, err := crclientcfg.GetConfig()
+	if err != nil {
+		return "", errors.Wrap(err, "unable to get config to talk to kubernetes api server")
+	}
+
+	client, err := clientset.NewForConfig(cfg)
+	if err != nil {
+		return "", errors.Wrap(err, "unable to get client from the given config")
+	}
+
+	host, err := client.ConfigV1().Infrastructures().Get(context.TODO(), "cluster", meta.GetOptions{})
+	if err != nil {
+		return "", errors.Wrap(err, "unable to get cluster infrastructure resource")
+	}
+	// get API server internal url of format https://api-int.abc.devcluster.openshift.com:6443
+	if host.Status.APIServerInternalURL == "" {
+		return "", errors.Wrap(err, "could not get host name for the kubernetes api server")
+	}
+	return host.Status.APIServerInternalURL, nil
 }
 
 // getWICDCredentials returns the CA cert and access token associated with the WICD service account
