+++ pkg/services/services.go
@@ -200,7 +200,7 @@ func generateKubeletArgs(argsFromIgnition map[string]string, debug bool) ([]stri
 		kubeletArgs = append(kubeletArgs, fmt.Sprintf("--%s=%s", ignition.CloudProviderOption, cloudProvider))
 	}
 	if cloudConfigValue, ok := argsFromIgnition[ignition.CloudConfigOption]; ok {
-		// cloud config is placed by WMCB in the c:\k directory with the same file name
+		// cloud config is placed by WMCO in the c:\k directory with the same file name
 		cloudConfigPath := windows.K8sDir + filepath.Base(cloudConfigValue)
 		kubeletArgs = append(kubeletArgs, fmt.Sprintf("--%s=%s", ignition.CloudConfigOption, cloudConfigPath))
 	}
